+++ index.js
@@ -37,14 +37,15 @@ const gitClientFactory   = require('./src/git.js');
 const github             = require('./src/github.js');
 const taoInstanceFactory = require('./src/taoInstance.js');
 
-const data          = {};
+const data          = {
+    wwwUser : 'www-data'
+};
 
 //TODO CLI params
 const origin        = 'origin';
 const baseBranch    = 'develop';
 const releaseBranch = 'master';
 const branchPrefix  = 'release';
-const wwwUser       = 'www-data';
 
 var taoInstance;
 var gitClient;
@@ -96,7 +97,7 @@ config.load()
         default: data.taoRoot || process.cwd()
     }) )
     .then( result => {
-        taoInstance = taoInstanceFactory(path.resolve(result.taoRoot), false, wwwUser);
+        taoInstance = taoInstanceFactory(path.resolve(result.taoRoot), false, data.wwwUser);
         return taoInstance.isRoot();
     })
     .then( result => {
@@ -263,10 +264,13 @@ config.load()
 
 
 // Update translations
+    .then( () => log.doing('Translations') )
+    .then( () => log.warn('Update translations during a release only if you know what you are doing') )
     .then( () => inquirer.prompt({
         type: 'confirm',
         name: 'translation',
         message: `${data.extension.name} needs updated translations ? `,
+        default : false
     }) )
     .then( result => {
         if(result.translation){